requirements.txt
asgiref==3.8.1
blinker==1.8.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
dnspython==2.6.1
filelock==3.16.1
Flask==3.0.3
identify==2.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nodeenv==1.9.1
platformdirs==4.3.6
pre-commit==3.8.0
PyYAML==6.0.2
virtualenv==20.26.6
Werkzeug==3.0.4
